refactor: split Header component into SignInMenu & SignOutMenu
