+++ frontend/src/components/Header/Header.jsx
@@ -1,131 +1,26 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
 
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import HeaderLogo from '@/assets/images/header-logo.svg';
-import { ReactComponent as CameraIcon } from '@/assets/images/camera-icon.svg';
 import { flexMixin } from '@/styles/mixins';
-import ProfileIcon from '@/assets/images/profile-icon.svg';
-import STATUS from '@/constants/statusCode';
-import { removeItemFromLocalStorage } from '@/util';
-import { AUTH_TOKEN_LIST } from '@/constants/auth';
 
 import { UserContext } from '@/store/UserStore';
-import { ModalContext } from '@/store/ModalStore';
-import { Button, Box, IconButton, DropDown } from '@/components/Common';
-import { fetchCreateChannel, fetchChannelOwnedByUser } from '@/apis/channel';
-import { fetchUpdateNickname } from '@/apis/user';
-import {
-    NicknameModalContent,
-    ChannelDetailModalContent,
-    LoginModalContent,
-    ChannelModalContent,
-} from '@/components/ModalContent';
+
+import SignInMenu from './SignInMenu';
+import SignOutMenu from './SignOutMenu';
 
 export default function Header() {
-    const { handleModal, openModal } = useContext(ModalContext);
-    const { userInfo, dispatch } = useContext(UserContext);
-
-    const navigate = useNavigate();
-
-    const changeNicknameHandler = () => {
-        handleModal(<NicknameModalContent 
-            userInfo={userInfo}
-            dispatch={dispatch}
-            onSubmit={handleOnChangeNickname} />);
-    };
-
-    const logoutHandler = () => {
-        removeItemFromLocalStorage(AUTH_TOKEN_LIST);
-        dispatch({ type: 'SIGN_OUT' });
-        navigate(window.location.pathname);
-    };
-
-    const handleOnClickCameraIcon = async () => {
-        const { user } = userInfo;
-        const { data: channelInfo, status } = await fetchChannelOwnedByUser(
-            user.id,
-        );
-
-        if (status === STATUS.NO_CONTENT) {
-            openModal(
-                <ChannelModalContent
-                    subHandleOnSubmit={handleOnCreateChannel}
-                    successText="방송 시작"
-                />,
-            );
-        } else if (status === STATUS.OK) {
-            openModal(<ChannelDetailModalContent channelInfo={channelInfo} />);
-        }
-    };
-
-    const handleOnCreateChannel = async formData => {
-        try {
-            const channelID = await fetchCreateChannel(formData);
-            navigate(`/stream-manager/${channelID}`);
-        } catch (err) {
-            throw new Error(err);
-        }
-    };
-
-    const handleOnChangeNickname = async data => {
-        try {
-            const response = await fetchUpdateNickname({
-                ...data,
-                id: userInfo.user.id,
-            });
-            return response;
-        } catch (err) {
-            throw new Error(err);
-        }
-    };
-
-    const profileDropDownItems = () => {
-        const items = [
-            ['닉네임 변경', changeNicknameHandler],
-            ['로그아웃', logoutHandler],
-        ];
-
-        const dropDownItems = items.map(([text, handler]) => {
-            return { text, handler };
-        });
-
-        return dropDownItems;
-    };
+    const {
+        userInfo: { isSignIn },
+    } = useContext(UserContext);
 
     return (
         <HeaderWrap>
             <Link to="/">
-                <Logo src={HeaderLogo} alt="header-logo" />
+                <Logo src={HeaderLogo} alt="header_logo" />
             </Link>
-
-            {!userInfo.isSignIn ? (
-                <Button
-                    text="로그인"
-                    size="medium"
-                    onClick={() => handleModal(<LoginModalContent />)}
-                />
-            ) : (
-                <Box>
-                    <IconButton
-                        size="large"
-                        type="square"
-                        onClick={handleOnClickCameraIcon}
-                    >
-                        <CameraIcon />
-                    </IconButton>
-
-                    <DropDown
-                        toggleButtonChild={
-                            <IconButton type="square" size="large">
-                                <Logo src={ProfileIcon} />
-                            </IconButton>
-                        }
-                        items={profileDropDownItems()}
-                        contentPos={{ top: '4rem', right: '0' }}
-                    />
-                </Box>
-            )}
+            {!isSignIn ? <SignInMenu /> : <SignOutMenu />}
         </HeaderWrap>
     );
 }
