+++ frontend/src/components/Header/SignInMenu/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SignInMenu';
