+++ frontend/src/components/Header/SignOutMenu/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SignOutMenu';
